servers/mcp-news-server/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-news-server"
version = "0.1.0"
description = "MCP stdio server: curate world news from RSS, web pages, and SearXNG with deduplication"
requires-python = ">=3.12"
dependencies = [
  "mcp>=1.24.0,<2",
  "httpx>=0.28.1,<0.29",
  "feedparser>=6.0.11,<7",
  "pyyaml>=6.0.1,<7",
]

[project.scripts]
mcp-news-server = "mcp_news_server.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mcp_news_server = ["py.typed", "default_feeds.yaml"]
